requirements.txt
Flask==2.3.3
psycopg2
python-dotenv==1.0.0
SQLAlchemy>=2.0.31
Flask-SQLAlchemy==3.0.5
flask-cors == 6.0.1